Convert model to a graph.

Only segments from an input model are extracted and used to build the graph.

# S3 method for SC
as.igraph(x, ..., layout = TRUE)

# S3 method for sf
as.igraph(x, ..., layout = TRUE)

sc_as_igraph(x, ..., layout = TRUE)

# S3 method for SC0
as.igraph(x, ..., layout = TRUE)

Arguments

x
input model

...
arguments passed to methods

layout
keep the input vertex coordinates x-y as the graph layout, defaults to `TRUE` otherwise no layout is provided ## sf ## spatstat ## raw track data ## make up segments starting with the unjoin thing, that might show the way forward

Examples

data("minimal_mesh", package = "silicate")
as.igraph(minimal_mesh)

#> IGRAPH 0a406d8 DN-- 14 15 -- 
#> + attr: name (v/c), x (v/n), y (v/n), edge_ (e/c)
#> + edges from 0a406d8 (vertex names):
#>  [1] CDqehQ->TgewRR TgewRR->qaYUiJ qaYUiJ->M6TLX0 8n9QBD->M6TLX0 8n9QBD->trck1g
#>  [6] Rd8jX9->trck1g CDqehQ->Rd8jX9 YrnMbd->DNvVcN DNvVcN->YSvzYp ZBFEBY->YSvzYp
#> [11] MqosX5->ZBFEBY YrnMbd->MqosX5 trck1g->8LK6gs 8LK6gs->tNghlO Rd8jX9->tNghlO